flip_bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -6,7 +6,7 @@
  * @m: another number
  * Return: number of bytes
  */
-unsigned long flip_bits(unsigned long int n, unsigned long int m)
+unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned long int xor_result = n ^ m;
 	unsigned int count = 0;
